requirements.txt
click==8.1.7
pandas==2.1.4
pydantic==2.5.3
rich==13.7.0
boto3==1.28.74
python-dotenv==1.0.0
pyyaml==6.0.1
